requirements.txt
cryptography>=41.0.0
pyotp>=2.8.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0